pyproject.toml
[project]
name = "project-sam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "albumentations>=2.0.8",
    "altair>=6.0.0",
    "click>=8.3.1",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "decord>=0.6.0",
    "einops>=0.8.2",
    "flatbuffers==25.12.19",
    "ftfy>=6.3.1",
    "humanfriendly==10.0",
    "iopath>=0.1.10",
    "ml-dtypes==0.5.4",
    "mpmath==1.3.0",
    "numpy>=2.2.6",
    "onnx==1.20.1",
    "onnxruntime-gpu==1.23.2",
    "onnxslim==0.1.84",
    "opencv-python-headless>=4.13.0.92",
    "packaging==26.0",
    "pillow>=12.1.1",
    "protobuf==6.33.5",
    "pycocotools>=2.0.11",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "regex>=2026.2.28",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "scipy>=1.15.3",
    "setuptools<71",
    "streamlit>=1.55.0",
    "sympy==1.14.0",
    "tenacity>=9.1.4",
    "tensorrt-cu12==10.14.1.48.post1",
    "tensorrt-cu12-bindings==10.14.1.48.post1",
    "tensorrt-cu12-libs==10.14.1.48.post1",
    "timm>=1.0.25",
    "toml>=0.10.2",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tornado>=6.5.4",
    "tqdm>=4.67.3",
    "typing-extensions==4.15.0",
    "ultralytics>=8.4.21",
    "watchdog>=6.0.0",
]
